requirements.txt
psycopg2==2.7.3.1
python-decouple==3.1
slackclient==1.0.9
textblob==0.13.0
tweepy==3.5.0
